Refactors SortClientListCommand
