+++ src/main/java/seedu/address/logic/parser/AddressBookParser.java
@@ -23,7 +23,7 @@ import seedu.address.logic.commands.RedoCommand;
 import seedu.address.logic.commands.ScheduleCommand;
 import seedu.address.logic.commands.SelectCommand;
 import seedu.address.logic.commands.UndoCommand;
-import seedu.address.logic.commands.exceptions.SortClientCommand;
+import seedu.address.logic.commands.SortClientCommand;
 import seedu.address.logic.parser.exceptions.ParseException;
 
 /**
